refactor: rename file "db_util" to "utils"
- rename file "db_util" to "utils"
- rename file "db_util.test" to "utils.test"
- rename variables "db_util" to "utils"

(rename to match other file names [camelCase])
